add some kind of brush selection
